Initial commit of known-stable code
This commit contains the initial work done on the client before
implementing the grpc service and fleshing out the public interfaces.
The code has been running for a month support a bot with 40 thousand
servers and so is considered stable. This commit is mainly for
posterity and archival. It is not meant to be used.